Moved Statistics function to a different file; Removed some dead code
